Built-in Functions, Refactoring
This commit adds basic support for built-in functions. Both the support for these functions and the functions themselves require some more work in the future.

This commit also continues the ongoing code refactoring efforts. As of this commit, all of interpret.c has been refactored.

Finally, the logo resources added in a previous commit have been moved to a new directory.