Polishing, small improvements
* Fixed: Missing error on redefinition of built-in function.
* Updated README.
* Removed make.cmd.
* Some other small changes.

Profiling

In other news, I did some profiling of Une in its state as of this commit to get an insight into which parts of the program could potentially need optimization at some point.
The results show that over 50% of total execution time is spent lexing. Close to half of this time is spent pulling characters from the input stream.
Optimizing this process could be approached in two ways:
* Using `mmap()` or similar to map the input files into memory.
* Not using streams.

As for the other major steps:
* Parsing requires about 12% of total execution time.
* Interpretation requires about 35% of total execution time.

As for the profiling itself:
* I modified the source to repeat `main` 100 000 times.
* I compiled with `UNE_DEBUG`, `UNE_DEBUG_REPORT` and `UNE_DEBUG_SIZES` disabled, at the highest optimization level, with gcc's `-pg` flag.
* The input script was `examples/arithmetic_interpreter.une`.
* I used gprof to convert `gmon.out` into a human-readable format.
* I used [gprof2dot](https://github.com/jrfonseca/gprof2dot) and [Graphviz](https://graphviz.org/) to create a graphical representation of the report.