Reworked built-in function implementation
This commit slighty changes the way built-in functions work.
Instead of each function having its own custom declaration, we use a common
declaration that receives a list of all arguments.

This commit also fixes a bug that was not entirely fixed in the previous
commit, where the underscore in the error display can be offset if the input
contains invisibile characters.