Implement method chaining
Objects now have an 'owner' field, which contains a pointer to the
context that is active during the object's creation (copying also counts
as creating). interpret_stmts() then only dereferences objects if they
are owned by the current context, while interpret_return() never
dereferences results anymore. This way, objects are passed around by
reference until they are passed *up* to a foreign context, enabling
method chaining.

One nonideal, but unavoidable detail with this solution is that, to
ensure object copies receive the correct owner, the .copy method for the
object datatype needs access to the interpreter state. Passing the
interpreter state via argument would have created a huge mess, so I
decided to use a global variable une_is. une_is is currently only used
for this purpose, but I will refactor the rest of the code to make use
of it in a future commit.