Fix syntax error caused by inconsistent parsing
When interpreting a function body, we ensure the
result is correctly dereferenced in the STMTS
interpreter. This logic is lost when the function
body isn't wrapped in a STMTS node. To avoid this,
we now force the parser to wrap function bodies in
a STMTS node, even if they only contain a single
statement.

This commit also improves the LOGPARSE macros.